# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: FeeDLooP API
  description: |
    FeeDLooP is a feedback collection service for software development companies. It provides APIs for managing projects, collecting feedback through embeddable widgets, and managing bug reports, initiatives, and feedback submissions.

    ## Authentication
    Most endpoints require authentication using NextAuth.js session cookies. The widget submission endpoint is public to allow embedding.

    ## Rate Limiting
    API endpoints implement rate limiting to prevent abuse. Refer to security documentation for specific limits.

    ## CORS
    Different CORS policies are applied based on endpoint:
    - Widget endpoints: Permissive CORS for embedding
    - Dashboard/API endpoints: Restrictive CORS for security

    ## Widget Integration
    The embeddable widget uses the `/api/widget/submit` endpoint to submit feedback from any website. See the Framework Integration Guide for implementation details.

  version: 1.0.0
  contact:
    name: FeeDLooP API Support
    url: https://github.com/your-org/feedloop
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:3000
    description: Development server
  - url: https://localhost:3001
    description: Alternative development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/projects/{id}:
    get:
      tags:
        - Projects
      summary: Get project details
      description: Returns detailed information about a specific project
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Projects
      summary: Update project
      description: Updates project information (owner only)
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: Project name
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Projects
      summary: Delete project
      description: Permanently deletes a project and all associated data (owner only)
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID
      responses:
        '204':
          description: Project deleted successfully
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    Project:
      type: object
      required:
        - id
        - name
        - owner_id
        - integration_key
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique project identifier
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Project name
        owner_id:
          type: string
          format: uuid
          description: User ID of project owner
        integration_key:
          type: string
          pattern: '^[a-zA-Z0-9]{16,32}$'
          description: Unique key for widget integration
        created_at:
          type: string
          format: date-time
          description: Project creation timestamp
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            updated_at:
              type: string
              format: date-time
              description: Last update timestamp
            team_members:
              type: array
              items:
                $ref: '#/components/schemas/TeamMember'
            report_counts:
              type: object
              properties:
                total:
                  type: integer
                active:
                  type: integer
                archived:
                  type: integer
                by_type:
                  type: object
                  properties:
                    bug:
                      type: integer
                    initiative:
                      type: integer
                    feedback:
                      type: integer
...
    TeamMember:
      type: object
      required:
        - user_id
        - email
        - role
        - joined_at
      properties:
        user_id:
          type: string
          format: uuid
          description: User identifier
        email:
          type: string
          format: email
          description: User email
        name:
          type: string
          nullable: true
          description: User full name
        role:
          type: string
          enum: [owner, admin, member]
          description: Role in project
        joined_at:
          type: string
          format: date-time
          description: Date joined project
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
              message:
                type: string
                description: Specific error message
          description: Detailed validation errors
        code:
          type: string
          description: Error code for programmatic handling
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
security:
  - sessionAuth: []
...